chore: make the db path configurable

+++ clear_db.cc
@@ -4,7 +4,7 @@
 #include <unistd.h>
 #include "storage.h"
 
-int main() {
+int main(int argc, char* argv[]) {
   if (const char* env_p = std::getenv("BUILD_WORKING_DIRECTORY")) {
     if (chdir(env_p) != 0) {
       std::cerr << "Failed to switch to workspace directory.\n";
@@ -12,6 +12,12 @@ int main() {
   }
 
   std::string db_path = "test_db";
+  if (argc > 1) db_path = argv[1];
+  if (db_path.empty()) {
+    std::cerr << "DB path cannot be empty.\n";
+    return 1;
+  }
+
   rocksdb::Options options;
   rocksdb::Status status = rocksdb::DestroyDB(db_path, options);
   std::cout << "Status of destroying DB: " << status.ok() << "\n";
